Be more careful about recording temp lemmas.
Whenever we generate an obligation with verifyOrCreateT,
send it directly to the shell/plugin via the KernelEnv.
Plugin records it as before, shell now checks to see if the
lemma can be proven in the top-level context. If so, it records
and puts on proof stack. If not, it puts on proof stack without
recording. So it has to be proved either way, but local lemmas
will not be recorded after proof, so they can't be used
unsafely.